fix compile warnings on 64-bit Snow Leopard

+++ types.h
@@ -3,7 +3,7 @@
 
 #include <inttypes.h>
 
-typedef uint32_t ref_t;
+typedef unsigned long ref_t;
 typedef struct env env_t;
 typedef ref_t (*fn_t)(env_t *env, ref_t func, ref_t args);
